syncs: add Map.RangeMutable to allow mutations while iterating
A sensible operation is to mutate the current map while iterating
perhaps by deleting the current entry. This is disallowed by Range,
which only provides a read-only iteration over the map.

Provide an alternative RangeMutable API that allows for mutations
while iterating over the map. We effectively implement a
reentrant mutex by passing a second *Map to the iteration function
that is specially marked as already having the lock held.

Adding a *Map argument to the iteration function provides an
explicit way to implement reentrant operations,
rather than depending on unsafe stack scanning.

Updates tailscale/corp#9115

Signed-off-by: Joe Tsai <joetsai@digital-static.net>